DesktopCube@yare: Make compatible with Cinnamon 3.8+ (#179)
* Use stage dimensions instead of the first monitor

This ensures the cube effect is centered and visible for multi-monitor
setups.

* Misc clean up

- Add eslintrc, fix warnings
- Minor formatting for readability
- Update deprecated Clutter API usage
- Use arrow functions where possible for older Spidermonkey engines

* Fix CJS GC errors

We need to make sure we cache and remove MetaBackgroundActor before its
parent gets GC'd.

(cinnamon:8845): Cjs-CRITICAL **: 00:35:00.250: Attempting to call back
into JSAPI during the sweeping phase of GC. This is most likely caused
by not destroying a Clutter actor or Gtk+ widget with ::destroy signals
connected, but can also be caused by using the destroy(), dispose(), or
remove() vfuncs. Because it would crash the application, it has been
blocked and the JS callback not invoked.

(cinnamon:8845): Cjs-CRITICAL **: 00:35:00.250: The offending signal was
destroy on MetaBackgroundActor 0x557b88b666f0.
== Stack trace for context 0x557b880ab4c0 ==

* More clean up

- Remove more deprecated API usage
- Use for loops instead of forEach, its faster
- Fix inconsistent style

* Add compatibility for 3.8, 4.0

* Remove unused property

* Fix typo

* Minor style tweak

This is easier for me to read.