Win7AltTab@entelechy: Misc. clean up (#180)
* Fix undefined label warning

* Add eslintrc, fix warnings

* Use a single window preview captured from MetaWindowActor

The typical approach of opacifying every window is CPU intensive, and
bad UX on multiple monitors. This uses the same method ITM uses for
overlaying a window preview.

* Fix invalid object access when switching workspaces

* Misc. clean up

- Update deprecated Clutter API usage
- Improve formatting
- Remove iteration of an array with for...in
- Clear the _checkDestroyedTimeoutId timeout
- Remove unnecessary uses of Lang.bind
- Use lexical scope for utility functions

* Add compatibility for 3.8, 4.0

* Replace remaining for...in loops iterating arrays

Fix ternary formatting

* Don't let window items get clipped outside the viewable container

This was not expanding past 3 rows as windows are added. This probably
needs to be wrapped in an StScrollView container as it will eventually
clip the monitor, but this makes it usable for most situations.

* Fix the title not updating on hover

* Remove reference to global.window_manager

* Remove circular references, clean up on disable