[Flipper@connerdev] Fix for opacity related xsession-errors messages
Fix xsession-errors messages when using the "Back" or "Elastic" easing options. The message would report out of range opacity values. This fixes the problem by tweening the opacity on its own and using a hard-coded easing which will ensure the opacity value can not go below or above the target defined in the tweener.