[smart-panel@mohammad-sn] Use Flipper/DesktopCube in more cases
Use a new API that will be added to Flipper & DesktopCube for switching the workspace. This API will allow switching to arbitrary workspaces rather than just to workspaces that are adjacent to the current one. The change checks if the new API exists (which it does not currently, but Flipper & DesktopCube updates will follow soon) and uses the API if it exists. If the API is not found, the code will use DesktopCube/Flipper as it did before (in a slightly more limited way then what is possible using the API).