FEATURE Added hotkey for AutoTileMainList
added a very hacky workaround to get an additional hotkey to trigger
auto tiling.
this was achieved directly in the JS rather than the TS as I wasn't sure
how to access the functionality via TS namespacing.
I was also unable to get the hotkey to register in the cinnamon shortcut
interface for configuration, would appreciate a code review to point out
necessary changes for merging.