BlurCinnamon@klangman: Fix panel background actor hierarchy
Previously the blur background was considered an actor of the global
overlay actor. This causes 2 problems:
1. The state of the program with regards to fullscreen windows becomes
harder to understand and manage, and makes it very easy to
introduce bugs to the extension from changing any panel related code.
Primary the issue that is likely to occur is that the background of the
panel will be drawn over the full screen window in the position of where
the panel would be, even though no other aspects/buttons/etc of the
panel are even there.
2. The "transparent background" is actually blocking all the windows
below it, making it awkward if you try to drag a window under the
panel, it will look like it disappears or that the panel is not actually
transparent but actually a foreground (because that's what the global
overlay actor actually is).

So the fix to make this more robust is to make the blur background a
child of the panel actor itself, instead of the global overlay actor.